#4 Add Month and year to the file name
There is no date taken on the tadpole files so it's impossible to tell when the picture was taken in google photos. Thus, to aid users to find out this information, append the month and year to the file name so at least others can click on the file to know.